web_scraper/requirements.txt
scrapy
scrapy-playwright
flask
flask-sqlalchemy
requests
beautifulsoup4
pandas
python-dotenv
gunicorn
playwright
readability-lxml 